+++ src/Disco/Typecheck/Erase.hs
@@ -1,4 +1,3 @@
-{-# LANGUAGE ViewPatterns #-}
 -----------------------------------------------------------------------------
 -- |
 -- Module      :  Disco.Typecheck.Erase
