+++ src/Disco/Types/Rules.hs
@@ -1,5 +1,3 @@
-{-# LANGUAGE DeriveGeneric         #-}
-{-# LANGUAGE MultiParamTypeClasses #-}
 -----------------------------------------------------------------------------
 -- |
 -- Module      :  Disco.Types.Rules
